+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [1.1.0] - 05/2023 - @bugsounet: Global review of the code
+ * replace request by axios
+ * use a seed request for don't use cache when download result
+ * force playing video on display new GIF
+ * move fetch interval to node_helper
+ * fetch all last posts (generaly 10)
+ * add rotateInterval for ramdom display
+
 ## [1.0.0] - Unreleased
 
 First public release